+++ src/AITT.cc
@@ -49,8 +49,8 @@ AITT::~AITT(void)
 {
 }
 
-void AITT::SetWillInfo(const std::string &topic, const void *data, const size_t datalen,
-      AittQoS qos, bool retain)
+void AITT::SetWillInfo(const std::string &topic, const void *data, const int datalen, AittQoS qos,
+      bool retain)
 {
     return pImpl->SetWillInfo(topic, data, datalen, qos, retain);
 }
@@ -71,23 +71,23 @@ void AITT::Disconnect(void)
     return pImpl->Disconnect();
 }
 
-void AITT::Publish(const std::string &topic, const void *data, const size_t datalen,
+void AITT::Publish(const std::string &topic, const void *data, const int datalen,
       AittProtocol protocols, AittQoS qos, bool retain)
 {
     if (AITT_PAYLOAD_MAX < datalen) {
-        ERR("Invalid Size(%zu)", datalen);
+        ERR("Invalid Size(%d)", datalen);
         throw std::runtime_error("Invalid Size");
     }
 
     return pImpl->Publish(topic, data, datalen, protocols, qos, retain);
 }
 
-int AITT::PublishWithReply(const std::string &topic, const void *data, const size_t datalen,
+int AITT::PublishWithReply(const std::string &topic, const void *data, const int datalen,
       AittProtocol protocol, AittQoS qos, bool retain, const SubscribeCallback &cb, void *cbdata,
       const std::string &correlation)
 {
     if (AITT_PAYLOAD_MAX < datalen) {
-        ERR("Invalid Size(%zu)", datalen);
+        ERR("Invalid Size(%d)", datalen);
         throw std::runtime_error("Invalid Size");
     }
 
@@ -95,12 +95,12 @@ int AITT::PublishWithReply(const std::string &topic, const void *data, const siz
           correlation);
 }
 
-int AITT::PublishWithReplySync(const std::string &topic, const void *data, const size_t datalen,
+int AITT::PublishWithReplySync(const std::string &topic, const void *data, const int datalen,
       AittProtocol protocol, AittQoS qos, bool retain, const SubscribeCallback &cb, void *cbdata,
       const std::string &correlation, int timeout_ms)
 {
     if (AITT_PAYLOAD_MAX < datalen) {
-        ERR("Invalid Size(%zu)", datalen);
+        ERR("Invalid Size(%d)", datalen);
         throw std::runtime_error("Invalid Size");
     }
 
@@ -119,10 +119,10 @@ void *AITT::Unsubscribe(AittSubscribeID handle)
     return pImpl->Unsubscribe(handle);
 }
 
-void AITT::SendReply(MSG *msg, const void *data, size_t datalen, bool end)
+void AITT::SendReply(MSG *msg, const void *data, int datalen, bool end)
 {
     if (AITT_PAYLOAD_MAX < datalen) {
-        ERR("Invalid Size(%zu)", datalen);
+        ERR("Invalid Size(%d)", datalen);
         throw std::runtime_error("Invalid Size");
     }
